+++ console/controllers/AgentsController.php
@@ -7,6 +7,7 @@
  */
 namespace skeeks\cms\shop\console\controllers;
 use skeeks\cms\shop\models\ShopFuser;
+use skeeks\cms\shop\models\ShopProductPriceChange;
 use yii\console\Controller;
 use yii\db\Exception;
 use yii\db\Expression;
@@ -20,6 +21,7 @@ use yii\helpers\Console;
  */
 class AgentsController extends Controller
 {
+
     /**
      * Просмотр созданных бекапов баз данных
      */