+++ messages/ru/app.php
@@ -385,6 +385,7 @@ return
     "Profiles buyer"                                            => "Профили покупателя",
     "All of the buyer profile"                                  => "Все профили покупателя",
     "All customer orders"                                       => "Все заказы покупателя",
+    "Removing the old price changes"                            => "Удаление старых изменений цен",
     'Functional "warehouses" is not set in your store. Create new content, and specify it in the general settings of your store as a storage content.'                                           => 'В вашем магазине не настроен функционал "склады". Создайте новый контент, и укажите его в общих настройках вашего магазина в качестве складского контента.',
     'In this section, you can customize what content you can sell on your site.'  => 'В этом разделе вы можете настроить какой контент можно продавать на вашем сайте.',
 ];